# openapi/openapi.yaml
openapi: 3.0.0
info:
  title: API
  version: 1.0.0
paths:
  /messages:
    get:
      summary: Get all messages
      tags:
        - messages
      responses:
        '200':
          description: A list of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Message'
    post:
      summary: Create a new message
      tags:
        - messages
      requestBody:
        description: The message to create
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Message'
      responses:
        '201':
          description: The created message
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
    put:
      summary: Update message text by ID
      tags:
        - messages
      parameters:
        - name: id
          in: query
          required: true
          schema:
            type: string
          description: ID of the message to update
      requestBody:
        description: Update for message text
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                text:
                  type: string
                  description: The updated text of the message
      responses:
        '200':
          description: Updated message
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '404':
          description: Message not found

    delete:
      summary: Delete existing message by ID
      tags:
        - messages
      parameters:
        - name: id
          in: query
          required: true
          schema:
            type: string
          description: ID of the message to delete
      responses:
        '204':
          description: Deleted message
        '404':
          description: Message not found

components:
  schemas:
    Message:
      type: object
      properties:
        id:
          type: integer
          format: uint
        text:
          type: string